requirements.txt
msal
aiosmtpd
python-dotenv